Refactor setting default exception handler to fileutils
