+++ src/java/org/apache/cassandra/service/CassandraDaemon.java
@@ -71,16 +71,11 @@ import org.apache.cassandra.db.*;
 import org.apache.cassandra.db.commitlog.CommitLog;
 import org.apache.cassandra.exceptions.ConfigurationException;
 import org.apache.cassandra.exceptions.StartupException;
-import org.apache.cassandra.io.FSError;
-import org.apache.cassandra.io.sstable.CorruptSSTableException;
-import org.apache.cassandra.io.ExceededDiskThresholdException;
 import org.apache.cassandra.io.util.FileUtils;
 import org.apache.cassandra.metrics.CassandraMetricsRegistry;
 import org.apache.cassandra.metrics.DefaultNameFactory;
-import org.apache.cassandra.metrics.StorageMetrics;
 import org.apache.cassandra.service.StorageServiceMBean.NonTransientError;
 import org.apache.cassandra.thrift.ThriftServer;
-import org.apache.cassandra.tracing.Tracing;
 import org.apache.cassandra.utils.*;
 
 /**
@@ -220,45 +215,7 @@ public class CassandraDaemon
 
         maybeInitJmx();
 
-        Thread.setDefaultUncaughtExceptionHandler(new Thread.UncaughtExceptionHandler()
-        {
-            public void uncaughtException(Thread t, Throwable e)
-            {
-                StorageMetrics.exceptions.inc();
-                logger.error("Exception in thread " + t, e);
-                Tracing.trace("Exception in thread {}", t, e);
-                for (Throwable e2 = e; e2 != null; e2 = e2.getCause())
-                {
-                    JVMStabilityInspector.inspectThrowable(e2);
-
-                    if (e2 instanceof FSError)
-                    {
-                        if (e2 != e) // make sure FSError gets logged exactly once.
-                            logger.error("Exception in thread " + t, e2);
-
-                        if (e2.getCause() instanceof CorruptSSTableException)
-                            FileUtils.handleCorruptSSTable((CorruptSSTableException) e2.getCause());
-                        else if (e2.getCause() instanceof ExceededDiskThresholdException)
-                            FileUtils.handleExceededDiskThreshold((ExceededDiskThresholdException) e2.getCause());
-                        else
-                            FileUtils.handleFSError((FSError) e2);
-                    }
-
-                    if (e2 instanceof CorruptSSTableException)
-                    {
-                        if (e2 != e)
-                            logger.error("Exception in thread " + t, e2);
-                        FileUtils.handleCorruptSSTable((CorruptSSTableException) e2);
-                    }
-                    if (e2 instanceof ExceededDiskThresholdException)
-                    {
-                        if (e2 != e)
-                            logger.error("Exception in thread " + t, e2);
-                        FileUtils.handleExceededDiskThreshold((ExceededDiskThresholdException) e2);
-                    }
-                }
-            }
-        });
+        FileUtils.setDefaultUncaughtExceptionHandler();
 
         Directories.startVerifyingDiskDoesNotExceedThreshold();
         completeSetupMayThrowSstableException();
